requirements.txt
langchain
chromadb
sentence-transformers
openai
tqdm
python-dotenv
langchain-community
google-generativeai
requests
streamlit
langchain-chroma
pyjwt
bcrypt
pytest
rank-bm25
scikit-learn
numpy
plotly
altair